Reuse OtherLogsPage directly
