+++ launcher/InstancePageProvider.h
@@ -46,7 +46,7 @@ class InstancePageProvider : protected QObject, public BasePageProvider {
         // values.append(new GameOptionsPage(onesix.get()));
         values.append(new ScreenshotsPage(FS::PathCombine(onesix->gameRoot(), "screenshots")));
         values.append(new InstanceSettingsPage(onesix));
-        values.append(new OtherLogsPage(inst));
+        values.append(new OtherLogsPage("logs", tr("Other logs"), "Other-Logs", inst));
         return values;
     }
