+++ launcher/ui/pages/instance/LogPage.h
@@ -42,11 +42,23 @@
 #include "BaseInstance.h"
 #include "launch/LaunchTask.h"
 #include "ui/pages/BasePage.h"
-#include "ui/pages/global/LauncherLogPage.h"
 
 namespace Ui {
 class LogPage;
 }
+class QTextCharFormat;
+
+class LogFormatProxyModel : public QIdentityProxyModel {
+   public:
+    LogFormatProxyModel(QObject* parent = nullptr) : QIdentityProxyModel(parent) {}
+    QVariant data(const QModelIndex& index, int role) const override;
+    QFont getFont() const { return m_font; }
+    void setFont(QFont font) { m_font = font; }
+    QModelIndex find(const QModelIndex& start, const QString& value, bool reverse) const;
+
+   private:
+    QFont m_font;
+};
 
 class LogPage : public QWidget, public BasePage {
     Q_OBJECT
