+++ launcher/ui/pages/instance/OtherLogsPage.ui
@@ -128,6 +128,16 @@
          <item row="1" column="0" colspan="5">
           <layout class="QHBoxLayout" name="horizontalLayout1">
            <item>
+            <widget class="QCheckBox" name="trackLogCheckbox">
+             <property name="text">
+              <string>Keep updating</string>
+             </property>
+             <property name="checked">
+              <bool>true</bool>
+             </property>
+            </widget>
+           </item>
+           <item>
             <widget class="QCheckBox" name="wrapCheckbox">
              <property name="text">
               <string>Wrap lines</string>
